Add the following updates:     - split/organize the code in different files;     - Add a function (Node) PropDuration to return a time.Duration based on a DURATION prop;     - Add functions (Node) DigProperty/DigParameter to dig out the string value of a nested prop/param     - Add a function ParseCalendar to parse the ical data from a string.
