+++ ical/calendar.go
@@ -1,0 +1,5 @@
+package ical
+
+type Calendar struct {
+	Items []CalendarItem
+}
