+++ ical/properties.go
@@ -1,0 +1,9 @@
+package ical
+
+const (
+	VCALENDAR = "VCALENDAR"
+	VEVENT  = "VEVENT"
+	DTSTART = "DTSTART"
+	DTEND   = "DTEND"
+	DURATION = "DURATION"
+)
